+++ CHANGES.md
@@ -1,6 +1,10 @@
 ## v0.6
 
+* Add `binomial(n, p)` to calculate a binomial distribution.
+* Add `beta(a, b)` to calculate a beta distribution.
+* Add `bernoulli(p)` to calculate a bernoulli distribution.
 * Add `geomean` to calculate the geometric mean.
+* Add `laplace` to calculate Laplace's Law of Succession. If `s` and `n` are passed, it will calculate `(s+1)/(n+2)`. If `s`, `time_passed`, and `time_remaining` are passed, it will use the [time invariant version](https://www.lesswrong.com/posts/wE7SK8w8AixqknArs/a-time-invariant-version-of-laplace-s-rule). Use `time_fixed=True` for fixed time periods and `time_fixed=False` (default) otherwise.
 * Add `p_to_odds` to convert probability to odds. Also `odds_to_p` to convert odds to probability.
 * Add `geomean_odds` to calculate the geometric mean of odds, converted to and from probabilities.